Fix 2 log lines, and fix a warning.
git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/branches/2.3@11496 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/netplay/nettypes.c
@@ -471,5 +471,5 @@ void NETtest()
 
 int NETgetSource()
 {
-	return NetMsg.source;
+	return (int) NetMsg.source;
 }
